backend/python-scripts/requirements.txt
torch
torchvision
transformers
pillow
numpy
python-dotenv
